requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
idna==3.3
python-dotenv==0.19.2
requests==2.26.0
typer==0.4.0
urllib3==1.26.7
